+++ dutch/app/lib/turn-state.js
@@ -66,6 +66,9 @@ function createTurnState(deps) {
     ) {
       [a.player.cards[a.index], b.player.cards[b.index]] = [b.player.cards[b.index], a.player.cards[a.index]];
       deps.moveSlotMemoryForAllBots(a.player.id, a.index, b.player.id, b.index, 'Jack swap');
+      if (deps.moveHumanKnowledgeForAllBots) {
+        deps.moveHumanKnowledgeForAllBots(a.player.id, a.index, b.player.id, b.index, actorId);
+      }
       if (deps.observeDecisionForAllBots) {
         deps.observeDecisionForAllBots(actorId, 'jack-target', { targetId: a.player.id });
         deps.observeDecisionForAllBots(actorId, 'jack-target', { targetId: b.player.id });